api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bcrypt==4.3.0
certifi==2025.4.26
click==8.1.8
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
pipdeptree==2.26.0
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-jose==3.4.0
python-multipart==0.0.20
rich==14.0.0
rsa==4.9
sentry-sdk==2.26.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinydb==4.8.2
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
